+++ src/panel/pets.ts
@@ -1,25 +1,118 @@
-import { PetSize } from "../common/types";
+import { PetColor, PetSize, PetType } from "../common/types";
 import { ISequenceTree } from "./sequences";
 import { IState, States, resolveState, HorizontalDirection, ChaseState, BallState, FrameResult, PetInstanceState, isStateAboveGround } from "./states";
+import { CAT_NAMES, DOG_NAMES, CRAB_NAMES, SNAKE_NAMES, CLIPPY_NAMES, TOTORO_NAMES, DUCK_NAMES } from "../common/names";
 
 export class InvalidStateException {
 
 }
 
+export class PetElement {
+    el: HTMLImageElement;
+    collision: HTMLDivElement;
+    pet: IPetType;
+    color: PetColor;
+    type: PetType;
+  
+    constructor(el: HTMLImageElement, collision: HTMLDivElement, pet: IPetType, color: PetColor, type: PetType){
+      this.el = el;
+      this.collision = collision;
+      this.pet = pet;
+      this.color = color;
+      this.type = type;
+    }
+  }
+
+export interface IPetCollection {
+    pets(): Array<PetElement>;
+    push(pet: PetElement): void;
+    reset(): void;
+    seekNewFriends(): string[];
+    locate(name: string): PetElement | undefined;
+}
+
+export class PetCollection implements IPetCollection {
+    _pets: Array<PetElement>;
+
+    constructor(){
+        this._pets = new Array(0);
+    }
+
+    pets() {
+        return this._pets;
+    }
+
+    push(pet: PetElement){
+        this._pets.push(pet);
+    }
+
+    reset(){
+        this._pets = [];
+    }
+
+    locate(name: string): PetElement | undefined {
+        return this._pets.find((collection, value, obj) => {
+            return collection.pet.name() === name;
+        });
+    }
+
+    seekNewFriends() : string[] { 
+        if (this._pets.length <= 1)
+            {return [];} // You can't be friends with yourself.
+        var messages = new Array<string>(0);
+        this._pets.forEach(petInCollection => {
+            if (petInCollection.pet.hasFriend())
+                {return;} // I already have a friend!
+            this._pets.forEach(potentialFriend => {
+                if (potentialFriend.pet.hasFriend())
+                    {return;} // Already has a friend. sorry.
+                if (!potentialFriend.pet.canChase())
+                    {return;} // Pet is busy doing something else.
+                if (potentialFriend.pet.left() > petInCollection.pet.left() &&
+                    potentialFriend.pet.left() < petInCollection.pet.left() + petInCollection.pet.width())
+                    {
+                        // We found a possible new friend..
+                        console.log(petInCollection.pet.name(), " wants to be friends with ", potentialFriend.pet.name(), ".");
+                        if (petInCollection.pet.makeFriendsWith(potentialFriend.pet))
+                        {
+                            messages.push(`${petInCollection.pet.name()} (${petInCollection.pet.emoji()}): I'm now friends ❤️ with ${potentialFriend.pet.name()} (${potentialFriend.pet.emoji()})`);
+                        }
+                    }
+            });
+        });
+        return messages;
+    }
+}
+
 export interface IPetType {
+    nextFrame(): void
+
+    // Special methods for actions
     canSwipe(): boolean
     canChase(): boolean
     swipe(): void
     chase(ballState: BallState, canvas: HTMLCanvasElement): void
-    nextFrame(): void
+
+    // State API
     getState(): PetInstanceState
     recoverState(state: PetInstanceState): void
+    recoverFriend(friend: IPetType): void
+
+    // Positioning
     bottom(): number;
     left(): number;
     positionBottom(bottom: number): void;
     positionLeft(left: number): void;
     width(): number;
     floor(): number;
+
+    // Friends API
+    name(): string;
+    emoji(): string;
+    hasFriend(): boolean;
+    friend(): IPetType;
+    makeFriendsWith(friend: IPetType): boolean;
+    isPlaying(): boolean;
 } 
 
 function calculateSpriteWidth(size: PetSize): number{
@@ -47,8 +140,10 @@ abstract class BasePetType implements IPetType {
     private _bottom: number;
     petRoot: string;
     _floor: number;
+    _friend: IPetType | undefined;
+    private _name: string;
 
-    constructor(spriteElement: HTMLImageElement, collisionElement: HTMLDivElement, size: PetSize, left: number, bottom: number, petRoot: string, floor: number){
+    constructor(spriteElement: HTMLImageElement, collisionElement: HTMLDivElement, size: PetSize, left: number, bottom: number, petRoot: string, floor: number, name: string){
         this.el = spriteElement;
         this.collision = collisionElement;
         this.petRoot = petRoot;
@@ -58,6 +153,8 @@ abstract class BasePetType implements IPetType {
         this.initSprite(size, left, bottom);
         this.currentStateEnum = this.sequence.startingState;
         this.currentState = resolveState(this.currentStateEnum, this);
+
+        this._name = name;
     }
 
     initSprite(petSize: PetSize, left: number, bottom: number) {
@@ -110,9 +207,17 @@ abstract class BasePetType implements IPetType {
         return {currentStateEnum: this.currentStateEnum};
     }
 
+    recoverFriend(friend: IPetType){
+        // Recover friends..
+        this._friend = friend;
+    }
+
     recoverState(state: PetInstanceState){
+        // TODO : Resolve a bug where if it was swiping before, it would fail
+        // because holdState is no longer valid.
         this.currentStateEnum = state.currentStateEnum!;
         this.currentState = resolveState(this.currentStateEnum, this);
+
         if (!isStateAboveGround(this.currentStateEnum)){
             // Reset the bottom of the sprite to the floor as the theme
             // has likely changed.
@@ -125,7 +230,7 @@ abstract class BasePetType implements IPetType {
     }
 
     canChase(){
-        return !isStateAboveGround(this.currentStateEnum);
+        return !isStateAboveGround(this.currentStateEnum) && this.currentStateEnum !== States.chase;
     }
 
     swipe() {
@@ -180,6 +285,17 @@ abstract class BasePetType implements IPetType {
             this.faceRight();
         }
         this.setAnimation(this.currentState.spriteLabel);
+
+        // What's my buddy doing?
+        if (this.hasFriend() && this.currentStateEnum !== States.chaseFriend){
+            if (this.friend().isPlaying() && !isStateAboveGround(this.currentStateEnum))
+            {
+                this.currentState = resolveState(States.chaseFriend, this);
+                this.currentStateEnum = States.chaseFriend;
+                return;
+            }
+        }
+
         var frameResult = this.currentState.nextFrame();
         if (frameResult === FrameResult.stateComplete)
         {
@@ -196,12 +312,42 @@ abstract class BasePetType implements IPetType {
             this.currentState = resolveState(nextState, this);
             this.currentStateEnum = nextState;
         } else if (frameResult === FrameResult.stateCancel){
-            if (this.currentStateEnum === States.chase) { // Currently the only one anyway
+            if (this.currentStateEnum === States.chase) {
+                var nextState = this.chooseNextState(States.idleWithBall);
+                this.currentState = resolveState(nextState, this);
+                this.currentStateEnum = nextState;
+            } else if (this.currentStateEnum === States.chaseFriend) {
                 var nextState = this.chooseNextState(States.idleWithBall);
                 this.currentState = resolveState(nextState, this);
                 this.currentStateEnum = nextState;
             }
         }
+    }
+
+    hasFriend() : boolean {
+        return this._friend !== undefined;
+    }
+
+    friend() : IPetType { 
+        return this._friend!;
+    }
+
+    name(): string {
+        return this._name;
+    }
+
+    makeFriendsWith(friend: IPetType): boolean {
+        this._friend = friend;
+        console.log(this.name(), ": I'm now friends ❤️ with ", friend.name());
+        return true;
+    }
+
+    isPlaying(): boolean {
+        return this.currentStateEnum === States.runRight || this.currentStateEnum === States.runLeft ;
+    }
+
+    emoji(): string { 
+        return "🐶";
     }
 }
 
@@ -252,6 +398,9 @@ export class Totoro extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "🐾";
+    }
 }
 export class Cat extends BasePetType {
     label = "cat";
@@ -272,11 +421,11 @@ export class Cat extends BasePetType {
             },
             {
                 state: States.walkLeft,
-                possibleNextStates: [States.sitIdle, States.climbWallLeft]
+                possibleNextStates: [States.sitIdle, States.climbWallLeft, States.walkRight, States.runRight]
             },
             {
                 state: States.runLeft,
-                possibleNextStates: [States.sitIdle, States.climbWallLeft]
+                possibleNextStates: [States.sitIdle, States.climbWallLeft, States.walkRight, States.runRight]
             },
             {
                 state: States.climbWallLeft,
@@ -304,6 +453,9 @@ export class Cat extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "🐱";
+    }
 }
 
 export class Dog extends BasePetType {
@@ -329,11 +481,11 @@ export class Dog extends BasePetType {
             },
             {
                 state: States.walkLeft,
-                possibleNextStates: [States.sitIdle, States.lie]
+                possibleNextStates: [States.sitIdle, States.lie, States.walkRight, States.runRight]
             },
             {
                 state: States.runLeft,
-                possibleNextStates: [States.sitIdle, States.lie]
+                possibleNextStates: [States.sitIdle, States.lie, States.walkRight, States.runRight]
             },
             {
                 state: States.chase,
@@ -345,6 +497,9 @@ export class Dog extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "🐶";
+    }
 }
 
 export class Snake extends BasePetType {
@@ -366,11 +521,11 @@ export class Snake extends BasePetType {
             },
             {
                 state: States.walkLeft,
-                possibleNextStates: [States.sitIdle]
+                possibleNextStates: [States.sitIdle, States.walkRight, States.runRight]
             },
             {
                 state: States.runLeft,
-                possibleNextStates: [States.sitIdle]
+                possibleNextStates: [States.sitIdle, States.walkRight, States.runRight]
             },
             {
                 state: States.chase,
@@ -382,6 +537,9 @@ export class Snake extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "🐍";
+    }
 }
 
 export class Clippy extends BasePetType {
@@ -419,6 +577,9 @@ export class Clippy extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "📎";
+    }
 }
 
 export class RubberDuck extends BasePetType {
@@ -456,6 +617,9 @@ export class RubberDuck extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "🐥";
+    }
 }
 
 export class Crab extends BasePetType {
@@ -493,32 +657,57 @@ export class Crab extends BasePetType {
             },
         ]
     };
+    emoji(): string { 
+        return "🦀";
+    }
 }
 
 export class InvalidPetException {
 }
 
-export function createPet(petType: string, el: HTMLImageElement, collision: HTMLDivElement, size: PetSize, left: number, bottom: number, petRoot: string, floor: number) : IPetType {
+function getPetName(collection: Map<number, string>, label: string, count: number) : string {
+    if (collection.has(count)){
+        return collection.get(count)!;
+    } else {
+        return label + count;
+    }
+}
+
+export function createPet(petType: string, el: HTMLImageElement, collision: HTMLDivElement, size: PetSize, left: number, bottom: number, petRoot: string, floor: number, name: string | undefined, count: number) : IPetType {
     if (petType === "totoro"){
-        return new Totoro(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(TOTORO_NAMES, PetType.totoro, count);}
+        return new Totoro(el, collision, size, left, bottom, petRoot, floor, name);
     }
     if (petType === "cat"){
-        return new Cat(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(CAT_NAMES, PetType.cat, count);}
+        return new Cat(el, collision, size, left, bottom, petRoot, floor, name);
     }
     else if (petType === "dog") {
-        return new Dog(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(DOG_NAMES, PetType.dog, count);}
+        return new Dog(el, collision, size, left, bottom, petRoot, floor, name);
     }
     else if (petType === "snake") {
-        return new Snake(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(SNAKE_NAMES, PetType.snake, count);}
+        return new Snake(el, collision, size, left, bottom, petRoot, floor, name);
     }
     else if (petType === "clippy") {
-        return new Clippy(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(CLIPPY_NAMES, PetType.clippy, count);}
+        return new Clippy(el, collision, size, left, bottom, petRoot, floor, name);
     }
     else if (petType === "crab") {
-        return new Crab(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(CRAB_NAMES, PetType.crab, count);}
+        return new Crab(el, collision, size, left, bottom, petRoot, floor, name);
     }
     else if (petType === "rubber-duck") {
-        return new RubberDuck(el, collision, size, left, bottom, petRoot, floor);
+        if (name === undefined)
+            {name = getPetName(DUCK_NAMES, PetType.rubberduck, count);}
+        return new RubberDuck(el, collision, size, left, bottom, petRoot, floor, name);
     }
     throw new InvalidPetException();
 }