+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "vscode-pets" extension will be documented in this file.
 
-## [1.5.1]
+## [1.6.0]
 
+- Pets will now become friends with other pets and chase after them when they're playing
+- Pets will remember who their friends are using saved state
+- Pets are assigned names and you're notified of new friendships
 - [bugfix] fixed the ball chases for themes with a high floor, like the castle
 
 ## [1.5.0]